refactor: issue 2,3 피드백 반영
