+++ src/types/mockData.type.ts
@@ -1,16 +1,11 @@
 export type MockType = {
   date: string;
   id: string;
-  value_area: value_area_Dataset;
-  value_bar: value_bar_Dataset;
+  value_area: datasetType;
+  value_bar: datasetType;
 };
 
-type value_area_Dataset = {
-  x: string;
-  y: number;
-};
-
-type value_bar_Dataset = {
+type datasetType = {
   x: string;
   y: number;
 };
